Factor out tiling functionality to tiles module
To allow a `Tiles` (or possibly `Grid`) class to be developed,
the tile grid and edge grid creating functions have been moved
to their own module, with most functionality moving cleanly.
Hoever, the test directory has changed considerably, as a
conftest file was required to allow sharing of fixtures.
directory has changed considerably.
